cec-compliance: test CEC 2.0 power transitions
The CEC 2.0 standard requires that devices broadcast power state
changes to avoid polling of the power state.

Test that this is indeed implemented.

Note that there is a bug in the linux cec framework that prevents
this test from working. It is fixed in this patch:

https://patchwork.linuxtv.org/patch/60444

But it will take some time before this appears in distro kernels.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>