cec-ctl: add --phys-addr-from-edid-poll
This polls the given EDID waiting for changes due to disconnect or
reconnect changes, and updates the physical address when it has
changed.

It's a hack for Pulse-Eight devices that need to be informed when
the EDID changes, but for which the kernel cannot know to which
HDMI output they are connected. Userspace can use this option to
provide that link. Eventually it should be possible to tell the
kernel to with which HDMI output a Pulse-Eight CEC adapter is
associated. But until then this will have to do.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>