libcecutil: drop CEC_MSG_ prefix when logging messages
The CEC_MSG_ prefix corresponds to the opcode name in cec.h,
but userspace doesn't care about that. So drop this prefix,
making the logging a bit cleaner without that prefix everywhere.

The only unexpected change is for unknown opcodes: CEC_MSG
was used for that, this is now UNKNOWN.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>