v4l2-compliance: improve USERPTR streaming tests
Verify that no memory before/after the actual buffer was
written. This helps testing flaky DMA engines that say they
can do USERPTR with contiguous or scatter-gather DMA, but in
reality they can't cope with the corner cases in the first
and last page of the malloc()ed data where you only write
part of a page.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>