v4l2-compliance: fix bug/improve testing filler
The test for the filler bytes at the end of a buffer had a bug
that caused it to fail.

It now also keeps track of the bytesused values to deal with
variable length buffers and checks that as well, although it
will only warn if post-max-bytesused values have been overwritten,
since that is allowed (if weird).

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>