cec-ctl: add 'repeats' suboption to --stress-test-power-cycle
Some displays require multiple attempts to wake up a display or
put it into standby. With the 'repeats' sub-option you can have the
test repeat the power toggle attempt. If it is necessary to use this
sub-option, then that indicates a problem with the display. But for
testing it is useful to be able to use this sub-option.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>