cec-ctl: added polls suboption to --stress-test-power-cycle
By default this stress test tried to poll for a power state change
for up to 50 seconds. This is too pessimistic and the default has
been changed to 30 seconds. In addition a new <polls> suboption
was added to change the number of polls, just like was done with
the --test-power-cycle option.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>