cec-ctl: handle ENONET errors
If a cec ioctl returns ENONET (i.e. the CEC adapter is in the
process of configuring itself), then wait 1 second and retry.

This error was seen during a stress test run, but it is really
not an error the first time you see this. If it happens again
after a second then something is wrong and it is OK to return
an error.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>