cec-ctl: set wakeup_la correctly and handle ENONET & EINVAL in stress tests
The wakeup_la was set inside the stress test loops instead of
setting it at the start: at the point the display is in standby and
the LA will either be a valid LA or it is invalid, indicating that
the HPD is low when in standby. In that case Image View On can be
called with LA Unregistered (15) since we don't know the LA.

Once the display wakes up and a valid LA is found, use that LA for
all future Image View On messages.

Some transmits can also return ENONET (either the HPD is low, or the
adapter is still configuring) or EINVAL (Image View On with LA 15 only)
that are now logged, but ignored. These errors can happen in this test.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>