cec-ctl: don't transmit IMAGE_VIEW_ON every second
If the HPD was low, then the power cycle stress tests would transmit
an IMAGE_VIEW_ON message every second. It shouldn't do this: the
--test-power-cycle already repeats IMAGE_VIEW_ON once if the first
attempt failed, and --stress-test-power-cycle now has the 'repeats'
sub-option if more attempts are needed. But for a properly designed
display multiple attempts should not be needed.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>