cec-ctl: improve wait_for_pwr_state
If, while waiting for Standby, the HPD is pulled low, then do not
immediately assume the device is in standby, instead wait for two
consecutive low HPDs before reporting that the device is in standby.

Also, if from the initial test it is clear that the device keeps
the HPD high while in standby, ignore low HPDs as a sign of the
device being in standby.

Background: most displays keep the HPD high when in standby, but some
pull it low to save energy. So a low HPD indicates that the display is
in standby for that second category.

However, often the HPD is pulled low for a short while when doing a
power state transition, so just detecting a low HPD is not enough to
decide a display is in standby. Instead wait for two consecutive low
HPDs before making that decision.

In addition, a low HPD doesn't indicate standby at all for the first
class of displays and should just be ignored as a standby indicator.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>