v4l2-ctl: vidcap: Add support for the CSC API
With the CSC API, capture drivers can allow userspace to request
to set the colorspace fields of the pixelformat. This patch adds
supports for this feature. when calling 'v4l2-ctl --set-fmt-video'
on capture devices userspace can ask to change the fields
'colorspace', 'ycbcr/hsv_enc', 'quantization' and 'xfer_func'
of the format if the driver supports it.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>