v4l2-compliance: disable cache flag test for kernels < 5.9
With the introduction of cache hint flags in kernel 5.9 a test
was added to check that the flags are cleared if the driver does
not support cache hint flags. But for older kernels these flags
were never cleared and this test would always fail.

Skip the test if the reported V4L2 version is < 5.9.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Tested-by: Dave Stevenson <dave.stevenson@raspberrypi.com>