cec-compliance: don't skip tests if phys addr is invalid
The compliance test appeared to be successful if the remote
physical address is invalid (e.g. because Give Physical Address
timed out) since all tests were just skipped in that case.

So drop the check for a valid physical address and instead the
system_info_phys_addr() test will fail with a critical error.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>