v4l2-ctl: set  total number of native DTDs to 1
The EDIDs used with v4l2-ctl --set-edid type=XXXX set the
"total number of native DTDs" field in the CTA-861 extension block to
0, meaning that the first Short Video Descriptor is the preferred timing.

However, this should be 1, meaning that the first DTD in block 0
is the preferred timing. This is more compatible and avoid a warning
from the edid-decode utility that checks for this.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>