v4l2-ctl: fix INTEGER64 support, add common_print_control()
Split off the printing of a control value into a separate
common_print_control() function.

While working on that I noticed that controls of type INTEGER64
were not handled correctly, so fix that.

Also the control names shown by the old code when printing the
control values was inconsistent: sometimes the name from QUERY_EXT_CTRL
was used, sometimes the ctrl_id2str[] name. Standardize on the latter.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>