configure.ac: Add --without-libudev option to avoid automagic dep
configure currently checks for and links against libudev
unconditionally, If this was unwanted and the library is removed then
this can leave v4l-utils broken. This is a problem for distributions,
especially Gentoo Linux where it affects end users. libudev is
unlikely to be removed entirely but the 32-bit library may be removed
from a 64-bit system, breaking a 32-bit build of v4l-utils.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>