cec-ctl: improve the --phys-addr-from-edid-poll option
When --phys-addr-from-edid-poll is specified, start a background
thread that does the polling. This avoids having to start one
cec-ctl instance to do EDID polling, and another to start a stress
test (--stress-test-power-cycle).

When all other requested actions are finished, then cec-ctl will just
wait for Ctrl-C when --phys-addr-from-edid-poll is specified, so this
effectively keeps the old behavior as well.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>