cec-ctl: don't spam log with PA changes unless paused
When a thread is started to poll for EDID changes, then that thread
was logging the new Physical Address whenever it changed, but that
messed up the logging of e.g. --stress-test-power-cycle. So now only
log PA changes if the main thread is paused since that's when nothing
else is happening.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>