dvbv5-zap: add a warning when not recording PMT data
There are several audio formats and at least one video
format that are sent using Elementary Stream type 6
(private data). The only way to identify the stream
format is by checking some descriptors that come together
with the PMT table, being the registration descriptor the
most important one, as it contains a FOURCC value that
should uniquelly identify the stream type.

Due to that, add a note when audio and/or video streams
are recorded without PMT data.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>