cec-ctl/cec-compliance: add pulse8-cec kernel version check
The CEC framework had problems with displays that pull down
the HPD when in Standby. This was fixed in kernel 5.5 and the fix
was backported to LTS kernels 4.19.94 and 5.4.9. Check that the right
kernel version is being used to prevent unreliable stress and
compliance test results.

For the stress test using the wrong kernel is a hard FAIL, for
the compliance test it is just a warning since it is much more
likely to pass.

This is only checked for the pulse8-cec driver, for other CEC devices
(most likely on an embedded system) you hopefully know what you are
doing.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>