cec-ctl: detect if Standby etc. messages are Nacked
During the power cycle stress test the <Active Source>, <Standby>
and <Image View On> messages are transmitted without checking if
the message was Nacked. Add this check and retry transmitting after
waiting for a second.

Some displays become unresponsive for a short time after changing
power state, and during that time these messages are NACKed. The
stress test didn't handle that very well, although it is dubious
for a display to do this, since this makes life hard for sources.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>