v4l2-compliance: test for vb2/m2m poll kernel bug
A bug has been discovered in the vb2 and m2m implementations of poll
where EPOLLIN and EPOLLOUT events are never signaled to epoll_wait()
unless at least one of them is included in the initial epoll_ctl() call.

The presence of this bug can be detected by setting an initial empty set
of events with EPOLL_CTL_ADD, then setting the events we are interested
in using EPOLL_CTL_MOD. epoll_wait() will then invariably timeout even
if there is a queue-related event pending.

Signed-off-by: Alexandre Courbot <gnurou@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>