Build with libbpf, remove local sources
libbpf is nowadays widely available in all distros that enable BPF,
and it is a standalone library with a release cycle and a stable API.
Remove the old sync of bpf.c/bpf.h and check for the library via
pkg-config and use it instead.
Only a minor modification is needed to make it compile.

Signed-off-by: Luca Boccassi <bluca@debian.org>
Signed-off-by: Sean Young <sean@mess.org>