libdvbv5: avoid a potential access out of an array
In practice, this condition is very unlikely to occur,
as it would depend on upgrading a Kernel with a newer DTV
property while using an older v4l-utils package.

As we don't add any new DTV property for quite a while,
I would say that only if someone would run a brand new
v4l2-utils and a very old Kernel (like 2.6.x), something
like that would happen.

In any case, better to be safe than sorry.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>