libdvbv5: dvb-dev-local: better handle realloc()
The realloc() call assumes that, when realloc() succeeds,
the return pointer is identical.

This should always be the case here, as the code is reducing
the size of the struct. Yet, an assumption like that can
be dangerous, if used with a library different than the
standard glibc.

So, let's store the returned pointer.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>