libdvb: fix ATSC service location parser
A fixup patch actually broke the parser, as it was copying data
to the wrong place.

Yet, while testing it, it was noticed some streams that have a
descriptor 0xa1 whose the language code is bogus:

	|           pcr PID               1539
	|\ elementary service - 3 elementaries
	|-  elementary 0
	|-      | stream type 0x00
	|-      | PID         1
	|-      | Language    .'. (0xcf 0x27 0x04)
	|-  elementary 1
	|-      | stream type 0x04
	|-      | PID         0
	|-      | Language    ..& (0x01 0xcf 0x26)
	|-  elementary 2
	|-      | stream type 0x05
	|-      | PID         1280
	|-      | Language    ... (0x00 0x01 0xcf)

So, clearly this is not an ATSC descriptor. So, improve the
logic to allow displaying invalid language codes, plus
check if the descriptor's size is what it was expected.

Fixes: 868ca548b10a ("libdvbv5: fix init of the ATSC service location")
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>