configure.ac: Fix building without libudev
Since the addition of the --without-libudev option automatic detection
of missing libudev have been broken and fails with,

    configure: error: Package requirements (libudev) were not met:

    Package 'libudev', required by 'virtual:world', not found

This is because autotools macro PKG_CHECK_MODULES action-if-not-found
behaves differently from most other macros,

    As most of the original autoconf macros, there are boolean values
    provided, for the cases when the check succeeded or failed. In
    contrast with almost all of the original macros, though, the default
    action-if-not-found will end the execution with an error for not
    having found the dependency.

As the AS_IF check following the PKG_CHECK_MODULES prints a warning if
libudev is not found the intention can't have been to fail configure if
libudev not found. Fix this by setting have_libudev=no instead of
failing due to the default action.

Fixes: 49316dc21f806d0c ("configure.ac: Add --without-libudev option to avoid automagic dep")
Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>