qv4l2: Add capture toggle and close hotkeys to CaptureWin
It can be inconvenient to stop or start the stream when the capture
window obscures most or all of the application window.  Register the
current application window stream toggle to the capture window.
Register a more standard Control-W to close the window.  Add both to
the context menu to allow them to be discovered.

Signed-off-by: David Fries <David@Fries.net>