cec-compliance: fix 'unresponsive' detection
The poll_stable_power_status() function reported the unresponsive_time,
which is not the duration the device is unresponsive during a power state
transition, but the time of the last unresponsive attempt since the start
of that state transition.

That's not what was intended, so change unresponsive_time to
unresponsive_cnt where we keep track of the number of unresponsive
attempts. Adjust the messages accordingly.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>