cec-compliance: wake up remote device if needed
Until now, if the remote device was in standby the test would just
abort with a message that the remote device should be woken up manually.

This patch changes this to attempt to automatically wake up the remote
device. Basically it adds the 'Wake up' to the core tests, but it is
only run if the remote device was in standby, otherwise it is skipped.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>