v4l2grab: optimize conversion routines
The conversion routines came from the vimc driver, where they
also re-scale and flip images vertically and horizontally
and can convert into several formats including Bayer.

So, it was done in two steps. As we just want to convert
everything into RGB24, this can be simplified in order
to avoid double-buffering.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>