v4l2grab: use BT.709 by default on YUV conversion
This is more commonly found those days. Yet, the right approach
would be, instead, to add full support for colorspace. So,
add a note there, while keeping the BT.601 table too.

As the hole idea of v4l2grab is to do quick tests, let's
keep using an integer arithimetic, as this should be faster.
Yet, let's improve its precision.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>