v4l2-compliance: relax time32-64 test
The delta between the 64 bit DQEVENT timestamp and the 32 bit timestamp
doesn't have to be <= 10 ms, that's too short when run in a VM instance.

Change it to 500 ms, that should be fine for this test. It's not about
performance, but about checking that the timestamps retrieved through
the 32 bit compat API are sane.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>