v4l2-compliance: improve select() check in captureBufs()
This test checks that at least one of the read or write fds are
set upon return from select(). But it is possible while streaming
that only an event is returned. In that case this check would fail.

Add a check for an event (exception) as well.

This should solve a regular occuring failure in the compliance test.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>