mc_nextgen_test: Display ancillary links
Add an argument to toggle display of ancillary links in the output of
mc_nextgen_test. Update media_show_links() to account for the extra
type of link and display them if configured by the right option.

The optargs table is reformatted, since with the new option name it
doesn't really fit in the original indentations.

Signed-off-by: Daniel Scally <djrscally@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>