cec-ctl: only generate eob for CEC pin events
An end-of-bit event is generated if a CEC pin event arrives that
indicates that the CEC bus remained high for longer than a bit period,
which means that it was the end of the message.

But this was also done for HPD/5V events, which is incorrect.
Especially since HPD/5V events might have a timestamp that is
later than a CEC event (probably due to a driver or framework
bug). In any case, the eob generator has nothing to do with HPD/5V
events.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>