cec-ctl: improve --analyze-pin performance
Analyzing large pin log files takes a long time, especially when combined
with --verbose and -w (wallclock time).

Parsing the file could be improved with some simple reordering of tests.

When generating the analyzer output the main bottleneck was logging the
timestamp, especially with -w. Optimizing this improved performance by
a factor of 3-4.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>