cec-ctl: store the wallclock/monotonic clocks every minute
In order to compensate for drift between the wallclock and monotonic
clock when logging, both clock timestamps are stored every hour when using
the --store-pin option. But that's not often enough since the drift
can be quite high (in the order of 1 ms per minute), so store this
every minute, not every hour.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>