qv4l2: fix search/replace mistake in vbi-tab.cpp/h license text
At some time in the past someone did search-and-replace of 'General'
to 'Vbi', but that also replaced "GNU General Public License" with
"GNU Vbi Public License". Obviously that was never intended, so
change that back.

It was there in the first commit of these sources back in 2012, so
I can't track it to a specific later commit.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Reported-by: Gregor Jasny <gjasny@googlemail.com>