cec-compliance: make <GET CEC VERSION> mandatory
Due to a confusing CEC 1.4 specification the GET CEC VERSION
message was optional. But this really makes no sense for modern
CEC implementations, and CEC 2.0 made it mandatory as well.

Without this message a source might fall back to CEC 1.3a, and
that's not what you want.

So make this mandatory in the compliance test.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>