test-media: check results of cmp in the vicodec tests
While the vicodec tests did verify if the output of several decode
tests was what was expected by using 'cmp', it did not report the
result in a 'FAIL/OK' format, so failures would not be detected
during the daily build.

Fix this to prevent undetected regressions in the future.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>