cec-ctl: --store-pin shouldn't enable pin monitoring
The use of --store-pin is meant to just set a filename for storing
the pin monitoring data, it is not meant to enable pin monitoring
itself, that's for --monitor-pin.

Using just --store-pin without --monitor-pin seemed to work, but it
was actually not logging some data, causing warnings to be issued
when analyzing the result.

With this patch --monitor-pin is required to actually enable pin
monitoring.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>