libv4lconvert: Don't ignore return value of ftruncate()
The ftruncate() function is declared with the warn_unused_result
attribute, which generates a warning when compiling libv4lconvert:

[60/340] Compiling C object lib/libv4lconvert/libv4lconvert.so.0.0.0.p/control_libv4lcontrol.c.o
../lib/libv4lconvert/control/libv4lcontrol.c: In function ‘v4lcontrol_create’:
../lib/libv4lconvert/control/libv4lcontrol.c:728:17: warning: ignoring return value of ‘ftruncate’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  728 |                 ftruncate(shm_fd, V4LCONTROL_SHM_SIZE);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Fix it by checking the return value and falling back to malloc-ed memory
for controls, as done when mmap() fails.

While at it, fix a typo in a comment, and drop an unneeded "error" word
from an error message.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>