meson: Use -include to include config.h instead of -I.
Use -include to get config.h included, instead of -I. and having the
source files include the config.h. The issue with the latter is that -I.
is rather late in the command line, and there may be config.h lurking in
some other project include dir (e.g. valgrind's).

For Android, add the -include option directly to the Android.mk files,
pointing to the android-config.h in the root directory.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
[Handle Android support]
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Tested-by: Sean Young <sean@mess.org>
Tested-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>