configure.ac: fix gconv directory search
By default the gconv libraries are installed in /usr/local/lib/gconv,
but the search for the default gconv system directory also checks that.

So the first time this will find the correct directory (/usr/lib/...),
but after installing gconv it will find /usr/local/lib/gconv and use
that as the link directory for libJIS.so, but that library isn't
there, so the build will fail.

Change the search order to check the /usr/local directories last.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>