build: do not fall back to libbpf header lookup
In case the first lookup fails due to an unmet version constraint
a lookup through the headers defeats that constraint and a too
old libbpf would be used.

To fix the issue, avoid the fallback and only use pkg-config files
for the lookup like the old autotools build system did.

Signed-off-by: Gregor Jasny <gjasny@googlemail.com>