cec-ctl: add --test-random-power-states stress test
The 'normal' CEC power state stress tests wait politely for
the display to reach the new state before switching to the
other state.

This test doesn't do this and just transmits Standby or
Image View On at random times. The device might not actually
switch to the other state, but it shouldn't lock up. After
10 state changes it will wait for the device to reach a
stable state, then it queries the power state. If this
fails, then the test fails.

Sadly, some displays can indeed lock up with this test...

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>