v4l2-tracer: add macros to print debug info
Instead of repeating the same code to print file, function and line info,
replace it with a macro so that it is easier to read and maintain. Use the
line_info() macro if the information should always be printed, otherwise
use debug_line_info() to print only when the debug option is set.

Use __FILE_NAME__ (if available) instead of __FILE__ to avoid printing the
entire relative path of the file location.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>