v4l2-tracer: stop stringifying v4l2_input "tuner"
The "tuner" field in struct v4l2_input is actually the tuner index used by
capture drivers that have more than one tuner. The comment in the UAPI
indicated that this field was an enum v4l2_tuner_type and so the
v4l2-tracer was (unsuccessfully) attempting to convert this field to a
string. Stop trying to stringify the "tuner" field and just let it be an
integer.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>