cec-compliance: get_power_status(): fix valid RX test
Explicitly test for a valid reply to the <Give Device Power Status>
message. The current test just checks that it isn't a FEATURE ABORT
or a time out, but that misses the ABORT status where the HPD goes
low while waiting for a reply.

Just check that the rx_status is OK instead.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>