v4l2-tracer: remove compress_frame_count
The trace context attempts to keep track of the number of compressed
frames received to avoid tracing the same decoded frame more than once.
However, this shared context only works if both the OUTPUT and CAPTURE
queues are run in the same process. If they are run in different
processes, then they will each have their own separate context and no
decoded frames will be traced.

Remove the compressed_frame_count since it is not functioning as expected.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>