v4l2-tracer: autogenerate files that can be added to git
A comment at the beginning of the autogenerated files provides an absolute
path to the location of the autogeneration script. In preparation for
adding these files to the git repository, remove this absolute path.

Also, the autogeneration script was adding an extra newline to
trace-gen.cpp which was causing a whitespace warning, so remove that
too.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>