v4l-utils: fix @PACKAGE_VERSION@ use
Man pages used @PACKAGE_VERSION@ which is supposed to be replaced by the
v4l-utils version (e.g. 1.27.0), instead it was replaced by "1.27.0", i.e.
including the double quotes. That messed up the man pages.

It's now replaced by @MAN_PACKAGE_VERSION@ with is the same version, but
without the double quotes.

This was also a problem in lib/include/libdvbv5/libdvb-version.h.in
where additional quotes where added. Those are now dropped.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>