media-ctl: Re-order setting format and routes
Currently media-ctl attempts to set formats that are passed to it
with -V _before_ setting routes passed to it with -R. This is a
problem, because the formats that one wants may not be valid until
routing has been configured (for example, if the format is for a
route that is inactive by default).

Reorder things so that setting routes comes before setting formats.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Reviewed-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Link: https://lore.kernel.org/r/20240117130805.939312-2-dan.scally@ideasonboard.com
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>