v4l2-compliance: Fix building against libc++
v4l2-test-time32-64.cpp included compiler.h, which checks
_LIBCPP_VERSION. This only works against libc++ when a C++ header has
already been included, which wasn't the case here.

The <version> header is the C++20 method of defining _LIBCPP_VERSION,
but for older versions, <ciso646> works as an alternative, so include
that in compiler.h.

compiler.h is for C as well as C++ though, so use __cplusplus to check
for a C++ compiler before including <ciso646>.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>