v4l2-ctl: replace --fix-edid-checksums by --keep-edid-checksums
The original idea was that v4l2-ctl could set EDIDs with checksum
errors, and that the option --fix-edid-checksums had to be given
to fix the checksums silently. But this was buggy and in fact
any EDID with checksum errors was rejected. It is however quite
useful to set an EDID with checksum errors for an HDMI receiver
in order to test handling of buggy EDIDs at the transmitter side.

This patch changes the code to always fix checksums silently
when setting a new EDID, unless the option --keep-edid-checksums
is given. In that case the EDID is used as-is.

This patch also fixes a bug when the EDID is dumped: that checks
for checksum errors and warns about that. But in two cases that
warning was mixed in with the output file instead of reported to
stderr.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>